Move SpeedLevelPreference enum class to data layer.
