+++ app/src/main/java/ar/com/westsoft/listening/data/datasource/Preferences.ext.kt
@@ -1,7 +1,7 @@
 package ar.com.westsoft.listening.data.datasource
 
 import androidx.datastore.preferences.core.Preferences
-import ar.com.westsoft.listening.domain.dictationgame.settings.SpeedLevelPreference
+import ar.com.westsoft.listening.screen.keyboard.ar.com.westsoft.listening.data.datasource.SpeedLevelPreference
 
 fun Preferences.getDictGameSettingsDSO() = DictGameSettingsDSO(
     readWordAfterCursor = this[PreferencesKey.ReadWordAfterCursor],
