+++ app/src/main/java/ar/com/westsoft/listening/data/datasource/SpeedLevelPreference.kt
@@ -1,4 +1,4 @@
-package ar.com.westsoft.listening.domain.dictationgame.settings
+package ar.com.westsoft.listening.screen.keyboard.ar.com.westsoft.listening.data.datasource
 
 enum class SpeedLevelPreference {
     LOW_SPEED_LEVEL,
