+++ app/src/main/java/ar/com/westsoft/listening/domain/dictationgame/settings/SetSpeedLevelUseCase.kt
@@ -1,6 +1,7 @@
 package ar.com.westsoft.listening.domain.dictationgame.settings
 
 import ar.com.westsoft.listening.data.repository.SettingsRepository
+import ar.com.westsoft.listening.screen.keyboard.ar.com.westsoft.listening.data.datasource.SpeedLevelPreference
 import javax.inject.Inject
 
 class SetSpeedLevelUseCase @Inject constructor(
