+++ app/src/main/java/ar/com/westsoft/listening/screen/dictationgame/game/DictGameMainScreen.kt
@@ -32,7 +32,7 @@ import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import ar.com.westsoft.listening.domain.dictationgame.settings.SpeedLevelPreference
+import ar.com.westsoft.listening.screen.keyboard.ar.com.westsoft.listening.data.datasource.SpeedLevelPreference
 import ar.com.westsoft.listening.screen.dictationgame.settings.DictGameSettingScreen
 import ar.com.westsoft.listening.screen.keyboard.MainKeyBoard
 import ar.com.westsoft.listening.screen.keyboard.ar.com.westsoft.listening.screen.dictationgame.game.GameConsoleScreen
