+++ app/src/main/java/ar/com/westsoft/listening/screen/dictationgame/settings/DictGameScreenSettingsState.kt
@@ -2,7 +2,7 @@ package ar.com.westsoft.listening.screen.dictationgame.settings
 
 import androidx.compose.ui.text.input.TextFieldValue
 import ar.com.westsoft.listening.data.repository.SettingsField
-import ar.com.westsoft.listening.domain.dictationgame.settings.SpeedLevelPreference
+import ar.com.westsoft.listening.screen.keyboard.ar.com.westsoft.listening.data.datasource.SpeedLevelPreference
 
 data class DictGameScreenSettingsState(
     val readWordAfterCursor: TextFieldValue,
