+++ app/src/main/java/ar/com/westsoft/listening/screen/dictationgame/settings/DictGameSettings.kt
@@ -6,7 +6,7 @@ import ar.com.westsoft.listening.data.datasource.PreferencesKey.ReadWordBeforeCu
 import ar.com.westsoft.listening.data.datasource.PreferencesKey.SpeechRatePercentage
 import ar.com.westsoft.listening.data.datasource.PreferencesKey.SpeedLevel
 import ar.com.westsoft.listening.data.repository.SettingsField
-import ar.com.westsoft.listening.domain.dictationgame.settings.SpeedLevelPreference
+import ar.com.westsoft.listening.screen.keyboard.ar.com.westsoft.listening.data.datasource.SpeedLevelPreference
 
 data class DictGameSettings(
     val readWordAfterCursor: SettingsField<Int>,
