+++ app/src/main/java/ar/com/westsoft/listening/screen/dictationgame/settings/DictGameSettingsScreen.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
-import ar.com.westsoft.listening.domain.dictationgame.settings.SpeedLevelPreference
+import ar.com.westsoft.listening.screen.keyboard.ar.com.westsoft.listening.data.datasource.SpeedLevelPreference
 import ar.com.westsoft.listening.screen.keyboard.ar.com.westsoft.listening.screen.dictationgame.settings.SelectableButton
 
 @OptIn(ExperimentalMaterial3Api::class)
