+++ app/src/main/java/ar/com/westsoft/listening/screen/dictationgame/settings/SelectableButton.kt
@@ -9,7 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import ar.com.westsoft.listening.data.repository.SettingsField
-import ar.com.westsoft.listening.domain.dictationgame.settings.SpeedLevelPreference
+import ar.com.westsoft.listening.screen.keyboard.ar.com.westsoft.listening.data.datasource.SpeedLevelPreference
 
 @Composable
 fun SelectableButton(
